// assignment3/github-profile-finder/src/interfaces.ts
export interface ISearchBarProps {
  getUser: (name: string) => Promise<void>;
}

export interface IUserInfo {
  login: string;
  id: number;
  node_id: string;
  avatar_url: string;
  gravatar_id: string;
  url: string;
  html_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  starred_url: string;
  subscriptions_url: string;
  organizations_url: string;
  repos_url: string;
  events_url: string;
  received_events_url: string;
  type: string;
  site_admin: boolean;
  name: string;
  company: string;
  blog: string;
  location: string;
  email: string | null;
  hireable: string | null;
  bio: string;
  twitter_username: string | null;
  public_repos: number;
  public_gists: number;
  followers: number;
  following: number;
  created_at: string;
  updated_at: string;
}

export interface IResultCardProps {
  data: IUserInfo | null;
}

export interface IUserState {
  status: string;
  data: IUserInfo | null;
}

export interface IResultProps {
  userState: IUserState;
}